ENH: moments were not calculated to the same precision after the removal of transforms in the image.

+++ Code/Common/itkImage.h
@@ -315,7 +315,8 @@ public:
     {
     for (unsigned int i = 0 ; i < VImageDimension ; i++)
       {
-      point[i] = static_cast<TCoordRep>(m_Origin[i] + m_Spacing[i] * index[i]);
+      point[i] = static_cast<TCoordRep>(m_Origin[i] + m_Spacing[i] * 
+        static_cast<double>( index[i] ));
       }
     }
